requirements.txt
django==3.2.5
ipdb
ipython
psycopg2_binary
requests
whitenoise
gunicorn
django-extensions
django-debug-toolbar==3.2.2
ldap3
mozilla_django_oidc
django-hijack
django-hijack-admin
boto3
ibm-cos-sdk
django-storages
pandas
django-sslserver
docutils
